Attempt to refactor text messages.
Worddle has a lot of logic dedicated to formatting text, and it makes the
code a little unreadable.  I tried to move most of the format strings out
as constants (and color them once instead of every time).  The resulting
code is a little cleaner looking, but I'm not sure I love it.