Attempt to make bountiful Worddle boards.
Generate 5 boards and pick the one with the most solutions.  It's more fun
when there are more solutions to be found.

Refactored a little - the board is now a class WorddleBoard.  Moved the
generation and solution discovery to WorddleBoard.  This simplifies Worddle a
little bit, but it probably could be simplified further.