Make ChatGPT base_url configurable (#73)
* Add configurable base_url

* Conditionally exclude a parameter if Gemini is the model

Sorry for the delay on this... The readme and help texts need updating, maybe few tweaks here and there, but obviously I never got around to it personally, and this is still a great addition, so going to just commit as is for now since you were thoughtful enough to make sure it doesn't break anything. Thanks again.